Move dot comparer to Elements.Search.
